DRY up embed code and fix inconsistencies between them
