+++ src/commands/addGame.ts
@@ -4,6 +4,7 @@ import { GamePlatform, Prisma } from "@prisma/client";
 
 import { Command, prisma, steam } from "..";
 
+import { gameDetailsEmbedBuilder } from "../util/embedTemplates";
 import { registerUser } from "../util/registerUser";
 
 interface SteamGameDetails {
@@ -129,17 +130,8 @@ export const addGame: Command = {
 
     await interaction.reply({
       embeds: [{
-        description: game.gameResource?.description ?? undefined,
-        fields: [
-          { inline: true, name: "Released", value: game.released ? "Yes" : "No" },
-          { inline: true, name: "Free?", value: game.free ? "Yes" : "No" },
-          { inline: true, name: "Players", value: `${game.minPlayers} - ${game.maxPlayers}` },
-        ],
+        ...gameDetailsEmbedBuilder(game),
         footer: { text: "Game successfully added." },
-        image: game.gameResource?.bannerImageURL ? { url: game.gameResource?.bannerImageURL } : undefined,
-        title: game.name,
-        thumbnail: game.gameResource?.thumbnailImageURL ? { url: game.gameResource?.thumbnailImageURL } : undefined,
-        url: game.gameURL ?? undefined,
       }],
     });
   },
